GA: Update legislators
Add four new legislators, as a result of two resignations and two deaths. There is additionally now a vacancy as a result of Jack Hill's death.

This also updates some names by removing nicknames and middle names, as these fail to match with the names found in bills and committees.